+++ matrix.h
@@ -3,15 +3,15 @@
 	#include <vector>
 
 	class matrix {
-		std::vector<std::vector<int>> pm;
+		std::vector<std::vector<double>> pm;
 		int num_rows, num_cols;
 
 		public:
 			matrix(int num_rows, int num_cols);
-			void set_row(int row_number, std::vector<int> row);
-			void set_col(int col_number, std::vector<int> col);
-			void set_val(int row_number, int col_number, int val);
-			void add_val(int row_number, int col_number, int val);
+			void set_row(int row_number, std::vector<double> row);
+			void set_col(int col_number, std::vector<double> col);
+			void set_val(int row_number, int col_number, double val);
+			void add_val(int row_number, int col_number, double val);
 			int get_val(int row_number, int col_number);
 			int get_rows();
 			int get_cols();
